safari-600.7-branch/LayoutTests/fast/forms/listbox-selection-expected.txt

<select> selection test for mouse events and keyevents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1) Select one item with mouse (no previous selection)
PASS selectionPattern("sl1") is "10000"
2) Select one item with mouse (with previous selection)
PASS selectionPattern("sl2") is "10000"
3) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl3") is "00001"
4) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl4") is "00100"
5) Attempt to select an item cmd-clicking
PASS selectionPattern("sl5") is "01000"
6) Attempt to select a range shift-clicking
PASS selectionPattern("sl6") is "01000"
7) Attempt to select a range with the keyboard
PASS selectionPattern("sl7") is "00100"
8) Select one item with mouse (no previous selection)
PASS selectionPattern("sl8") is "10000"
9) Select one item with mouse (with previous selection)
PASS selectionPattern("sl9") is "10000"
10) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl10") is "00001"
11) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl11") is "00100"
12) Select an item cmd-clicking
PASS selectionPattern("sl12") is "01100"
13) Select a range shift-clicking
PASS selectionPattern("sl13") is "11110"
14) Select a range with the keyboard
PASS selectionPattern("sl14") is "01100"
PASS successfullyParsed is true

TEST COMPLETE
